chore: Move ConversationSystemMessageCellDescription into its own file
